cli/requirements.txt
numpy>=1.26.0
tqdm>=4.66.3
Pillow>=10.3.0
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.2
scikit-learn>=1.3.0
matplotlib>=3.7.1
